Diploma/requirements.txt

pytest==8.4.1
allure-pytest==2.15.0

# === UI-тесты ===
selenium==4.35.0
webdriver-manager==4.0.2  # автоматическая загрузка ChromeDriver

# === API-тесты ===
requests==2.32.5

# === Работа с .env ===
python-dotenv==1.1.1

# === Линтер ===
autopep8              2.3.2